requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==4.0.1
django-bootstrap4==21.2
django-stdimage==5.3.0
Pillow==9.0.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
